backend/requirement.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-dotenv
langchain
langchain-community
langchain-openai
openai